requirements.txt
dotenv
requests
python-dateutil
influxdb
